apps/api/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pydantic==2.0.3
alembic==1.11.1
strawberry-graphql==0.192.0
python-dotenv==1.0.0
psycopg2-binary
